[New] Added test cases for pages admin builder components translations form
